JPA support test infrastructure

+++ sportsclub/sportsclub-hibernate-dao/src/main/java/org/jboss/snowdrop/samples/sportsclub/dao/hibernate/HibernateMembershipRepository.java
@@ -25,7 +25,7 @@ public class HibernateMembershipRepository extends HibernateRepository<Membershi
       return query.list();
    }
 
-   public int countAll()
+   public long countAll()
    {
       return (Integer)getCurrentSession().createCriteria(Membership.class).setProjection(Projections.count("code")).uniqueResult();
    }
